Separated DriveViewModel extensions into different files.
